Moves parameterized_frame to extensions
